Extract error/warning/hint/info messages directly from attributes
Background probes can now be asked to 'extract-report' from the
attributes that they analyse.  If so, they check if the attributes
are a collection of objects with a method

  public String getCodeProberReportString();

and, if so, calls that method on all elements of the collection.

This allows gathering "squiggle" marks without the need for side
effects.