Allow tagging protocol spec fields as EnumString
EnumString(options = {"foo", "bar"}) is intended to convey that a
string field may only take a value out of a predefined set of such
values.  The code generates no checks to that effect (at the moment)
but the information is available when reflecting on the fields of
a protocol.data.* class.